Added toStrings for spark dashboard
The spark dashboard on port 4040 includes beam metrics. Unfortunately, the `toString()` methods of relevant classes is not defined, so it's not possible to actually see the metrics that are being collected, as it simply prints the class name and a hash.

Added the required `toString` methods to show the data. Formatting can be improved, but at least it's there to see now.