Mixing beam code with spark
The Spark ML lib provides common ML algorithms. However, these do not work as a step in the pipeline, rather they take input RDDs and then create their own pipeline on top of that and run it. To make use of these algorithms in our code, a new runner is provided.

`NativeSparkRunner` allows to create a two stage pipeline:

1. the first stage is a regular beam pipeline;
2. the second stage can access the `RDD`s that have been created to implement the first phase and use native spark methods (including those from ML lib) on them.

It is not possible to combine beam and spark in any other way (e.g. creating `PCollection`s from the `RDD`s again and continuing the pipeline). However, it is of course possible to create a new pipeline that continue on from the first.

`LDATest` shows an example of how this can be used.